+++ lib/teemo.dart
@@ -1,2 +1,3 @@
 export 'src/teemo_impl.dart' show Teemo;
+export 'src/runes.dart' show Rune;
 export 'src/EventSubscription.dart' show EventSubscription;